[Qt][WK2] plugin tests assert in debug https://bugs.webkit.org/show_bug.cgi?id=106659
Reviewed by Benjamin Poulain.

Fixed the misuse of String::split.

* Shared/Plugins/Netscape/x11/NetscapePluginModuleX11.cpp:
(WebKit::truncateToSingleLine):
Add an assert to make sure this function works well. I just walked there
while debugging the bug, and although it was innocent, adding this assertion
is logical.
* UIProcess/Plugins/qt/PluginProcessProxyQt.cpp:
(WebKit::PluginProcessProxy::scanPlugin):
The ASSERT was wrong. We have three lines, each ends in a newline, so split
with allowEmptyEntries=true will devide it to 4 substrings.

Identifier: 125009@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@139600 268f45cc-cd09-0410-ab3c-d52691b4dbfc