[EFL][jhbuild] Bump the EFL dependencies in jhbuild to 1.7.5. https://bugs.webkit.org/show_bug.cgi?id=106574
Reviewed by Laszlo Gombos.

Rely on a more recent stable release, which includes support for
Lua 5.2 in Edje and other assorted minor fixes.

While here, also use the "configure" script to build the libraries
instead of relying on autogen.sh, which is not really needed with
a release tarball and may have problems with automake 1.13.

* efl/jhbuild.modules:

Identifier: 125015@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@139608 268f45cc-cd09-0410-ab3c-d52691b4dbfc