[BlackBerry] Zoom is still possible when user-scalable=no https://bugs.webkit.org/show_bug.cgi?id=106809
Patch by Jacky Jiang <zhajiang@rim.com>.
Reviewed by George Staikos.

PR: 274566
On wordpress.com, the contents were rendered at the width 343 which was
1 css pixel larger than the expected device-width 342 so that we didn't
respect the viewport.
There was a RenderBlock on the page which contained "\n" and an
inline-block. "\n" took 0.78 width and the inline-block took 342 width
and they were unfortunately placed in the same InlineBox which caused
the contents width 1 css pixels larger than the device-width.
Allow the contents width to be 1 css pixel larger and still respect the
viewport in such a special case.

* Api/WebPage.cpp:
(BlackBerry::WebKit::WebPagePrivate::respectViewport):

Identifier: 125034@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@139629 268f45cc-cd09-0410-ab3c-d52691b4dbfc