[BlackBerry] FCC appears when switching search engine providers https://bugs.webkit.org/show_bug.cgi?id=106824
Magnifying glass appears when switching search engine providers

PR 279178

Patch by Otto Derek Cheung <otcheung@rim.com> on 2013-01-14
Reviewed by Rob Buis.
Internal Reviewed by Mike Fenton.

We need to make sure fatfinger is touching a textfield when we tell notifyCaretChanged
to show the FCC.

This behaviour is accidentally regressed in the fix in PR 267177, causing the FCC to show
when the user touches an element that is not a textfield and changes the caret in the inputbar..

Tested that the behaviour in 267177 isn't reproducible and the FCC isn't invoked when we switch
search engine providers.

* WebKitSupport/SelectionHandler.cpp:
(BlackBerry::WebKit::SelectionHandler::caretPositionChanged):

Identifier: 125057@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@139667 268f45cc-cd09-0410-ab3c-d52691b4dbfc