Fix the Mac build sans ENABLE(CSS_FILTERS)
Reviewed by Darin Adler.

Include FloatRect in RemoteGraphicsLayer.mm; it's included by chance currently,
via something that must be compiled out without CSS_FILTERS (or SVG, or shaders, or something).

* WebProcess/WebPage/mac/RemoteGraphicsLayer.mm:

Move an export from a USE(ACCELERATED_COMPOSITING) block to a ENABLE(CSS_FILTERS)
block, to match the code that backs it.

* WebCore.exp.in:


Identifier: 125374@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140001 268f45cc-cd09-0410-ab3c-d52691b4dbfc