CanvasRenderingContext2D::drawTextInternal should create a mask buffer compatible with that of the canvas https://bugs.webkit.org/show_bug.cgi?id=107372
Reviewed by Simon Fraser.

No new tests, just matching acceleration modes.

* html/canvas/CanvasRenderingContext2D.cpp:
(WebCore::CanvasRenderingContext2D::drawTextInternal): Create a buffer with the same acceleration mode as the canvas
instead of forcing the use of an accelerated buffer if we support acceleration; the canvas could be unaccelerated
even if USE(IOSURFACE_CANVAS_BACKING_STORE) is on.


Identifier: 125617@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140263 268f45cc-cd09-0410-ab3c-d52691b4dbfc