[GTK] Enable CSS Image Orientation feature https://bugs.webkit.org/show_bug.cgi?id=99040
Patch by Zan Dobersek <zandobersek@gmail.com> on 2013-01-22
Reviewed by Martin Robinson.

Tools:

Enable the CSS Image Orientation feature in the development
builds of the GTK port. The feature is still disabled in the release
builds.

* Scripts/webkitperl/FeatureList.pm:

LayoutTests:

Remove the general expectation, replacing it with a failure expectation
for the test that's failing due to whitespace problems in the cssText attribute
value after the image orientation attribute is set.

* platform/gtk/TestExpectations:


Identifier: 125806@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140460 268f45cc-cd09-0410-ab3c-d52691b4dbfc