Invalidated SVG shadow tree should be always detached. https://bugs.webkit.org/show_bug.cgi?id=107634
Reviewed by Ryosuke Niwa.

Source/WebCore:

SVGUseElement::clearResourceReferences() uses removeAllChildren() for
clearing its shadow DOM, but this is wrong.
removeAllChildren() is designed for removing children of an out-of-document Node.
For efficiency, it skips a series of cleanup sequences like detach().

For removing SVG shadow tree which is in Document, removeChildren() should be used.
It does proper cleanup for the chilren.

Test: svg/custom/use-invalidate-click-crash.xhtml

* svg/SVGUseElement.cpp:
(WebCore::SVGUseElement::clearResourceReferences):

LayoutTests:

* svg/custom/use-invalidate-click-crash-expected.txt: Added.
* svg/custom/use-invalidate-click-crash.xhtml: Added.


Identifier: 125864@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140520 268f45cc-cd09-0410-ab3c-d52691b4dbfc