Source/WebCore: Avoid creating background layers on pages with a fixed background, but no image https://bugs.webkit.org/show_bug.cgi?id=107783 <rdar://problem/13074450>
Reviewed by Beth Dakin.

http://www.nme.com has background-attachment: fixed on the <body>, but
not background image. In that case there's no point making a layer
for the fixed root background.

Test: platform/mac/tiled-drawing/fixed-background/fixed-background-no-image.html

* rendering/style/RenderStyle.cpp:
(WebCore::allLayersAreFixed): Check to see if we have an image, as well
as fixed attachment.

LayoutTests: Avoid creating background layers on pages with a fixed background, but no image
https://bugs.webkit.org/show_bug.cgi?id=107783

Reviewed by Beth Dakin.

Testcase with body { background-attachment:fixed; } but no image.

* platform/mac/tiled-drawing/fixed-background/fixed-background-no-image-expected.txt: Added.
* platform/mac/tiled-drawing/fixed-background/fixed-background-no-image.html: Added.

Identifier: 125979@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140648 268f45cc-cd09-0410-ab3c-d52691b4dbfc